Change master template to force cache invalidation of partials
Caching problems should have been removed in T113095 but it's possible
that older wikis may continue to cache older values

Changing the master template is guaranteed to force a cache invalidation
so let's try this first. The safest way to do this without any visible
change is to add an HTML comment and is a technique we use in Minerva
to workaround this issue.

Bug: T248010
Change-Id: I7be7bbc6109a212962e12c08513830b5581e6c86